+++ test/macroy_web/controllers/org_file_controller_test.exs
@@ -2,33 +2,33 @@ defmodule  MacroyWeb.OrgFileControllerTest do
   use MacroyWeb.ConnCase
   alias Macroy.{OrgFile, Repo}
 
-  @blorp_org %OrgFile{
+  @blorp_org %{
     host: "localhost",
     path: "/home/blorp",
     filename: "crime.org"
   }
 
-  @css_org %OrgFile{
+  @css_org %{
     host: "localhost",
     path: "/home/css",
     filename: "communism.org"
   }
 
   setup do
-    Macroy.insert_org_file(@blorp_org)
-    Macroy.insert_org_file(@css_org)
+    Repo.insert(OrgFile.changeset(%OrgFile{}, @blorp_org))
+    Repo.insert(OrgFile.changeset(%OrgFile{}, @css_org))
   end
 
   describe "index/2" do
-    test "all files load" do
-      conn = get(conn, "/orgfiles")
-
+    test "all files load", %{conn: conn} do
       # Page doesn't explode in flames
-      resp = html_response(conn, 200)
+      resp = conn
+      |> get("/orgfiles")
+      |> html_response(200)
      
       # Assert that each orgfile shows up 
-      assert resp =~ @blorp.path
-      assert resp =~ @ccs_org.path
+      assert resp =~ @blorp_org.path
+      assert resp =~ @css_org.path
     end
   end
 end